Renamed LwEvent to Event class
