+++ app/src/main/kotlin/com/exactpro/th2/lwdataprovider/entities/responses/Event.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021-2024 Exactpro (Exactpro Systems Limited)
+ * Copyright 2021-2025 Exactpro (Exactpro Systems Limited)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,118 +17,95 @@
 package com.exactpro.th2.lwdataprovider.entities.responses
 
 import com.exactpro.cradle.messages.StoredMessageId
-import com.exactpro.th2.common.event.EventUtils
+import com.exactpro.cradle.testevents.StoredTestEventId
+import com.exactpro.cradle.testevents.lw.LwStoredTestEventSingle
+import com.exactpro.th2.common.event.EventUtils.toEventID
 import com.exactpro.th2.common.grpc.EventID
 import com.exactpro.th2.common.grpc.EventStatus.FAILED
 import com.exactpro.th2.common.grpc.EventStatus.SUCCESS
 import com.exactpro.th2.common.grpc.MessageID
 import com.exactpro.th2.common.message.toTimestamp
 import com.exactpro.th2.dataprovider.lw.grpc.EventResponse
-import com.exactpro.th2.lwdataprovider.entities.internal.ProviderEventId
 import com.exactpro.th2.lwdataprovider.grpc.toGrpcMessageId
 import com.google.protobuf.UnsafeByteOperations
-import io.javalin.openapi.OpenApiNullable
-import io.javalin.openapi.OpenApiPropertyType
-import io.javalin.openapi.OpenApiRequired
-import java.nio.ByteBuffer
-import java.time.Instant
+import com.exactpro.th2.common.grpc.Event as CommonGrpcEvent
 
 data class Event(
-    //** full event id  */
-    val eventId: String,
-    @get:OpenApiRequired
-    @get:OpenApiPropertyType(definedBy = String::class)
-    val batchId: String?,
-    /** last part of event id. it doesn't consider in equal and hashCode methods */
-    val shortEventId: String,
-    val isBatched: Boolean,
-    val eventName: String,
-    @get:OpenApiRequired
-    @get:OpenApiPropertyType(definedBy = String::class)
-    val eventType: String?,
-    @get:OpenApiRequired
-    @get:OpenApiPropertyType(definedBy = Instant::class)
-    val endTimestamp: Instant?,
-    val startTimestamp: Instant,
-    @get:OpenApiRequired
-    @get:OpenApiPropertyType(definedBy = String::class)
-    val parentEventId: ProviderEventId?,
-    val successful: Boolean,
-    val bookId: String,
-    val scope: String,
-    val attachedMessageIds: Set<String>,
-
-    @get:OpenApiRequired
-    @get:OpenApiNullable(nullable = false)
-    @get:OpenApiPropertyType(definedBy = Array<Any>::class)
-    val body: ByteBuffer?
+    val event: LwStoredTestEventSingle,
+    val batchId: StoredTestEventId? = null,
+    val parentBatchId: StoredTestEventId? = null,
 ) {
+    val eventId: StoredTestEventId
+        get() = event.id
 
-    fun convertToGrpcEventData(): EventResponse {
+    val bookId: String
+        get() = event.id.bookId.name
+
+    val scope: String
+        get() = event.id.scope
+
+    val attachedMessageIds: Set<StoredMessageId>
+        get() = event.messages ?: emptySet()
+
+    val isBatched: Boolean
+        get() = batchId != null
+
+    fun toGrpcEventResponse(): EventResponse = with(event) {
         return EventResponse.newBuilder()
-            .setEventId(EventUtils.toEventID(startTimestamp, bookId, scope, shortEventId))
+            .setEventId(id.toEventIdProto())
             .setIsBatched(isBatched)
-            .setEventName(eventName)
+            .setEventName(name)
             .setStartTimestamp(startTimestamp.toTimestamp())
-            .setStatus(if (successful) SUCCESS else FAILED)
-            .addAllAttachedMessageId(convertMessageIdToProto(attachedMessageIds))
+            .setStatus(if (isSuccess) SUCCESS else FAILED)
             .also { builder ->
-                batchId?.let { builder.setBatchId(EventID.newBuilder().setId(it)) }
-                eventType?.let { builder.setEventType(it) }
-                endTimestamp?.let { builder.setEndTimestamp(it.toTimestamp()) }
-                parentEventId?.let { builder.parentEventId = convertToEventIdProto(it) }
-                body?.let { builder.body = UnsafeByteOperations.unsafeWrap(body) }
+                batchId?.let { builder.batchId = it.toEventIdProto() }
+                parentId?.let { builder.parentEventId = it.toEventIdProto() }
+                type?.let { builder.eventType = it }
+                endTimestamp?.let { builder.endTimestamp = it.toTimestamp() }
+                attachedMessageIds.forEach {
+                    builder.addAttachedMessageId(it.toGrpcMessageId())
+                }
+                content?.let { builder.body = UnsafeByteOperations.unsafeWrap(it) }
             }.build()
+    }
+
+    fun toGrpcEvent(): CommonGrpcEvent = with(event) {
+        return CommonGrpcEvent.newBuilder()
+            .setId(id.toEventIdProto())
+            .setName(name)
+            .setType(type)
+            .setStatus(if (isSuccess) SUCCESS else FAILED)
+            .also { builder ->
+                parentId?.also { builder.parentId = it.toEventIdProto() }
+                type?.let { builder.type = it }
+                endTimestamp?.also { builder.endTimestamp = it.toTimestamp() }
+                attachedMessageIds.forEach {
+                    builder.addAttachedMessageIds(it.toGrpcMessageId())
+                }
+                content?.let { builder.body = UnsafeByteOperations.unsafeWrap(it) }
+            }
+            .build()
     }
 
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
-        if (javaClass != other?.javaClass) return false
+        if (other !is com.exactpro.th2.lwdataprovider.entities.responses.Event) return false
 
-        other as Event
-
-        if (eventId != other.eventId) return false
         if (batchId != other.batchId) return false
-        if (isBatched != other.isBatched) return false
-        if (eventName != other.eventName) return false
-        if (eventType != other.eventType) return false
-        if (endTimestamp != other.endTimestamp) return false
-        if (startTimestamp != other.startTimestamp) return false
-        if (parentEventId != other.parentEventId) return false
-        if (successful != other.successful) return false
-        if (bookId != other.bookId) return false
-        if (scope != other.scope) return false
-        if (attachedMessageIds != other.attachedMessageIds) return false
-        if (body != other.body) return false
+        if (event != other.event) return false
 
         return true
     }
 
     override fun hashCode(): Int {
-        var result = eventId.hashCode()
-        result = 31 * result + (batchId?.hashCode() ?: 0)
-        result = 31 * result + isBatched.hashCode()
-        result = 31 * result + eventName.hashCode()
-        result = 31 * result + (eventType?.hashCode() ?: 0)
-        result = 31 * result + (endTimestamp?.hashCode() ?: 0)
-        result = 31 * result + startTimestamp.hashCode()
-        result = 31 * result + (parentEventId?.hashCode() ?: 0)
-        result = 31 * result + successful.hashCode()
-        result = 31 * result + bookId.hashCode()
-        result = 31 * result + scope.hashCode()
-        result = 31 * result + attachedMessageIds.hashCode()
-        result = 31 * result + body.hashCode()
+        var result = batchId?.hashCode() ?: 0
+        result = 31 * result + event.hashCode()
         return result
     }
 
     companion object {
-        @JvmStatic
-        fun convertToEventIdProto(providerEventId: ProviderEventId): EventID = providerEventId.eventId.run {
-            EventUtils.toEventID(startTimestamp, bookId.name, scope, id)
-        }
-        @JvmStatic
-        fun convertMessageIdToProto(attachedMessageIds: Set<String>): List<MessageID> = attachedMessageIds.map { id ->
-            StoredMessageId.fromString(id).toGrpcMessageId()
-        }
+        fun StoredTestEventId.toEventIdProto(): EventID = toEventID(startTimestamp, bookId.name, scope, id)
+
+        fun Set<StoredMessageId>.toMessageIdProto(): List<MessageID> = map { id -> id.toGrpcMessageId() }
     }
 }
