+++ app/src/main/kotlin/com/exactpro/th2/lwdataprovider/entities/responses/LastScannedObjectInfo.kt
@@ -1,5 +1,5 @@
-/*******************************************************************************
- * Copyright 2021-2021 Exactpro (Exactpro Systems Limited)
+/*
+ * Copyright 2021-2025 Exactpro (Exactpro Systems Limited)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -12,7 +12,7 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- ******************************************************************************/
+ */
 
 package com.exactpro.th2.lwdataprovider.entities.responses
 
@@ -22,10 +22,6 @@ data class LastScannedObjectInfo(var id: String = "", var timestamp: Long = 0, v
 
     fun update(id: String, count: Long) {
         update(id, System.currentTimeMillis(), count)
-    }
-
-    fun update(event: BaseEventEntity, scanCnt: Long) {
-        update(event.fullEventId.toString(), event.startTimestamp, scanCnt)
     }
 
     fun update(message: ProviderMessage, scanCnt: Long) {
