+++ app/src/main/kotlin/com/exactpro/th2/lwdataprovider/http/TaskDownloadHandler.kt
@@ -31,7 +31,7 @@ import com.exactpro.th2.lwdataprovider.entities.requests.ProviderMessageStream
 import com.exactpro.th2.lwdataprovider.entities.requests.SearchDirection
 import com.exactpro.th2.lwdataprovider.entities.requests.SseEventSearchRequest
 import com.exactpro.th2.lwdataprovider.entities.responses.ser.HeapBufferPool
-import com.exactpro.th2.lwdataprovider.entities.responses.LwEvent
+import com.exactpro.th2.lwdataprovider.entities.responses.Event
 import com.exactpro.th2.lwdataprovider.entities.responses.ProviderMessage53
 import com.exactpro.th2.lwdataprovider.filter.FilterRequest
 import com.exactpro.th2.lwdataprovider.filter.events.EventsFilterFactory
@@ -246,7 +246,7 @@ class TaskDownloadHandler(
                         is EventTaskInfo -> {
                             val handler = HttpGenericResponseHandler(
                                 queue, responseBuilder, { it.run() }, dataMeasurement,
-                                LwEvent::eventId,
+                                Event::eventId,
                                 SseResponseBuilder::build
                             )
                             if (!taskInfo.attachHandler(handler)) return@execute TaskState.AlreadyInProgress
@@ -367,7 +367,7 @@ class TaskDownloadHandler(
         ) : TaskState()
         data class EventsReady(
             val info: EventTaskInfo,
-            val handler: HttpGenericResponseHandler<LwEvent>,
+            val handler: HttpGenericResponseHandler<Event>,
             val queue: ArrayBlockingQueue<Supplier<SseEvent>>,
         ) : TaskState()
     }
